# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.AccountResponse:
    properties:
      email:
        example: ada.new@example.com
        type: string
      username:
        example: ada_l
        type: string
    type: object
  dto.AccountUpdateRequest:
    properties:
      email:
        example: ada.new@example.com
        maxLength: 256
        type: string
      password:
        example: EvenStrongerP@ss1
        maxLength: 129
        minLength: 8
        type: string
      username:
        example: ada_l
        maxLength: 55
        minLength: 5
        type: string
    type: object
  dto.ConsentRequest:
    properties:
      given:
        example: true
        type: boolean
      type:
        example: privacy_policy
        type: string
      version:
        example: 2025-01
        type: string
    required:
    - given
    - type
    - version
    type: object
  dto.ConsentResponse:
    properties:
      created_at:
        example: "2025-01-03T10:00:00Z"
        type: string
      given:
        example: true
        type: boolean
      id:
        example: 11
        type: integer
      type:
        example: privacy_policy
        type: string
      updated_at:
        example: "2025-06-10T09:00:00Z"
        type: string
      user_id:
        example: 42
        type: integer
      version:
        example: 2025-01
        type: string
    type: object
...
  dto.DeleteConsentRequest:
    properties:
      type:
        example: privacy_policy
        type: string
      version:
        example: 2025-01
        type: string
    required:
    - type
    - version
    type: object
...
  dto.ListUserResponse:
    properties:
      total:
        example: 123
        type: integer
      users:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
    type: object
  dto.LoginRequest:
    properties:
      password:
        example: StrongP@ssw0rd
        maxLength: 256
        minLength: 8
        type: string
      username:
        example: ada_lovelace
        maxLength: 256
        type: string
    required:
    - password
    - username
    type: object
  dto.MeResponse:
    properties:
      email:
        example: ada@example.com
        type: string
      is_verified:
        example: true
        type: boolean
      roles:
        items:
          $ref: '#/definitions/dto.RoleResponse'
        type: array
      username:
        example: ada_lovelace
        type: string
    type: object
  dto.MessageResponse:
    properties:
      message:
        example: Operation successful
        type: string
    type: object
...
  dto.ProfileCreateRequest:
    properties:
      age:
        example: 28
        maximum: 150
        minimum: 16
        type: integer
      height:
        description: mm
        example: 1720
        maximum: 5000
        minimum: 200
        type: integer
      sex:
        example: female
        type: string
      weight:
        description: grams
        example: 65000
        maximum: 500000
        minimum: 20000
        type: integer
    type: object
  dto.ProfileResponse:
    properties:
      age:
        example: 28
        type: integer
      created_at:
        example: "2025-01-02T15:04:05Z"
        type: string
      height:
        example: 1720
        type: integer
      sex:
        example: female
        type: string
      updated_at:
        example: "2025-09-20T12:34:56Z"
        type: string
      weight:
        example: 65000
        type: integer
    type: object
  dto.ProfileUpdateRequest:
    properties:
      age:
        example: 29
        maximum: 150
        minimum: 16
        type: integer
      height:
        example: 1730
        maximum: 5000
        minimum: 200
        type: integer
      sex:
        example: female
        type: string
      weight:
        example: 64000
        maximum: 500000
        minimum: 20000
        type: integer
    type: object
...
  dto.RoleResponse:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: admin
        type: string
    type: object
...
  dto.TokenResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
info:
  contact:
    email: dev@example.com
    name: API Support
    url: https://example.com/support
  description: Public HTTP API for the Fitness Tracker app.
  title: Fitness Tracker API
  version: "1.0"
paths:
...
  /admin/users:
    get:
      consumes:
      - application/json
      description: Returns a paginated list of users with optional filtering and sorting.
      parameters:
      - description: Free-text search on username/email
        in: query
        name: q
        type: string
      - default: 1
        description: Page number (1-based)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        maximum: 200
        minimum: 1
        name: page_size
        type: integer
      - description: Sort field
        enum:
        - id
        - username
        - email
        - created_at
        - last_seen_at
        - is_verified
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: sort_dir
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: List users (admin)
      tags:
      - admin
  /admin/users/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Delete user (admin)
      tags:
      - admin
...
  /users/accounts:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AccountUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Update account fields
      tags:
      - users
  /users/consents:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Consent to delete
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteConsentRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Delete consent
      tags:
      - users
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ConsentResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: List consents
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: Consent payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ConsentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ConsentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Create consent
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ConsentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ConsentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Update consent
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Returns an access token in JSON and sets a refresh token cookie.
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Set-Cookie:
              description: refresh_token cookie (HttpOnly)
              type: string
          schema:
            $ref: '#/definitions/dto.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      summary: Login with username/password
      tags:
      - users
...
  /users/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Get current user info
      tags:
      - users
  /users/profile:
    delete:
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Delete profile
      tags:
      - users
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProfileResponse'
        "204":
          description: No Content when profile is empty
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "404":
          description: Profile not found
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Get profile
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: Profile payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ProfileCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Create profile
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ProfileUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Update profile
      tags:
      - users
  /users/refresh:
    post:
      description: Reads refresh token cookie and returns a new access token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TokenResponse'
        "401":
          description: Missing or invalid refresh token
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      summary: Refresh access token
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      summary: Register a new user
      tags:
      - users
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"